Fix add module POS

+++ app/Packages/Pos/Migrations/2026_04_07_100006_create_orders_table.php
@@ -8,16 +8,22 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('orders', function (Blueprint $table) {
-            $table->id();
-            $table->string('dni', 20)->nullable();
-            $table->json('detail')->nullable();
-            $table->decimal('total', 10, 2);
-            $table->foreignId('operator_id')->constrained('users')->onDelete('cascade');
-            $table->foreignId('status_id')->constrained('status_orders')->onDelete('cascade');
-            $table->boolean('paid')->default(false);
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('orders')) {
+            Schema::create('orders', function (Blueprint $table) {
+                $table->id();
+                $table->string('dni', 20)->nullable();
+                $table->json('detail')->nullable();
+                $table->decimal('total', 10, 2);
+                $table->foreignId('operator_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('status_id')->constrained('status_orders')->onDelete('cascade');
+                $table->string('sync_id', 36)->nullable()->unique();
+                $table->boolean('paid')->default(false);
+                $table->string('mp_payment_id', 50)->nullable();
+                $table->decimal('mp_transaction_amount', 10, 2)->nullable();
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     public function down(): void
